Added number of favorites in RowLayout

+++ app/src/main/java/mr/cooker/mrcooker/data/entities/Recipe.kt
@@ -8,7 +8,7 @@ data class Recipe(
     var timeToCook: Int = -1,
     var ingredients: String = "",
     var instructions: String = "",
-    var numOfFavorites: Int = -1,
+    var numOfFavorites: Long = -1,
     var showToEveryone: Boolean = true,
     var timePosted: Long = -1,
     var ownerID: String? = null,
